refactor: move shared cryptographic helpers into new ObliviousXHelpers module
